task-scheduling/requirements.txt
celery==5.4.0
python-dotenv==1.0.1
redis==5.0.8